+++ src/resolve.rs
@@ -287,7 +287,7 @@ impl ResolveContext {
     ) -> Result<bool> {
         let expr = Expression::parse(expr).map_err(Error::new)?;
         let mut cfg_map = self.cfg.borrow_mut();
-        cfg_map.eval_cfg(&expr, target, || self.rustc(build_config).clone().into())
+        cfg_map.eval_cfg(&expr, target, || self.rustc(build_config).into())
     }
 }